rename svgs => SocialIcons
