+++ packages/nextjs/app/builders/0xc26816404F551a7b3964B0D665380e9e3aca17b3/page.tsx
@@ -4,7 +4,7 @@ import Image from "next/image";
 import Link from "next/link";
 import { blo } from "blo";
 import { useEnsAvatar, useEnsName } from "wagmi";
-import { GithubIcon, LinkedInIcon, MailIcon, MapMarkerIcon, XIcon } from "~~/components/svgs";
+import { GithubIcon, LinkedInIcon, MailIcon, MapMarkerIcon, XIcon } from "~~/components/SocialIcons";
 
 export default function BatchPage() {
   const address = "0xc26816404F551a7b3964B0D665380e9e3aca17b3";
